requirements.txt
openai>=1.0
httpx
pyyaml
markdown
jinja2
pygments
beautifulsoup4
